// src/api/generated/getMeResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Cinema Platform API
 * API Gateway for Cinema Platform microservices
 * OpenAPI spec version: 1.0
 */
import type { GetMeResponseRole } from './getMeResponseRole';

export interface GetMeResponse {
  id: string;
  name?: string;
  email: string;
  phone: string;
  avatar?: string;
  role: GetMeResponseRole;
}